Interpreter: implemented WHILE/WEND, REPEAT/UNTIL, HEX$, multi DIM, SWAP, all maths functions, RANDOMIZE

+++ core/interpreter/interpreter.h
@@ -107,7 +107,7 @@ void itp_freeProgram(struct Core *core);
 
 union Value *itp_readVariable(struct Core *core, enum ValueType *type, enum ErrorCode *errorCode);
 struct TypedValue itp_evaluateExpression(struct Core *core, enum TypeClass typeClass);
-struct TypedValue itp_evaluateNumericExpression(struct Core *core, float min, float max);
+struct TypedValue itp_evaluateNumericExpression(struct Core *core, int min, int max);
 int itp_isEndOfCommand(struct Interpreter *interpreter);
 enum ErrorCode itp_endOfCommand(struct Interpreter *interpreter);
